Move call-buddy arch building to new build-arch.sh script and arch.txt
With the launchpad utility the idea is that any box connected via SSH
can instantly run call-buddy, regardless of whether it's installed on
that box or regardless of the architecture. Because we want to run on
most, if not all common architectures, this requires that architecture
specific binaries be compiled and installed ahead of time, lest we want
to require users to always have a go compiler handy at run time. We
currently do this in the telephono-ui/Makefile, but if a user or
developer wants to compile to a non-standard architecture, it's not
obvious nor documented how this would be done.

To fix this, the build-arch.sh script was created to basically enumerate
though the architectures defined in arch.txt. Having this present in the
root directory also makes it easier and more obvious to users/developers
on how to add a architecture, and correspondingly, documentation has
been written on how to do so.