Resolve user defined prefix in Makefile to abspath
If you were to do "make install prefix=../usr/local", the result would
be an error, rather than install to ../usr/local since that location
would later be interpreted as relative to launchpad/, telephono-ui/ and
telephono/, rather than the base directory when those Makefiles are
called.

The fix for this is to simply make the given prefix="..." variable an
absolute path so the prefix path is the same when in subdirectories.