Store view properties in map, use gocui to figure out current view
This commit overhauls the way we deal with the current view state in
an attempt to clean up the uglyness of the old approach of having two
different variables, one a iota and the other a string, in order to
switch and change view properties.

Now all the views are stored in a single global map that be accessed.
Updating the view is as simple as calling methods belonging to a struct
within that map.