Replace internal http.Request/Response structs with custom ones
This allows us to have better control over the body, which normally is
a io.ReaderCloser which presents problems when trying to read things
more than once. This problem prevented PR #37 (which allowed for hinting
of history entries) from working.

It also enables us to create custom methods on those new structs such as
the String() method in the response, further separating the logic
between the TUI and the store.