Depend on fixed Gui.Update() semantics in forked gocui
Basically gocui's Gui.Update() semantics made it so you couldn't
guarantee that a Update() with another Update() following would be
exeucted first which kinda sucks. This made our program occasionally
crash because we would change the manager in one Update() and then
attempt to change the view in another Update() which would crash since
the Manager hadn't yet been set!

@Jabbawookiez and me (@dylngg) literally spent a hour or so debugging
this problem before finding this out, it's not obvious for Go novices
like myself. So I forked gocui and fixed it.

The fix is explained here:
https://github.com/call-buddy/gocui/commit/f47c77447ba49144ffb95788a9bd40388e0ee67a

It was tricky to come up with that fix and get it implemented right.
It's definitely a contender for one of my most difficult-to-implement
programs with a small code size ratio. For the record, it was tested by
building call-buddy with the attempted fix and spamming ctrl-y at
unreasonable rates-that technique actually caught bugs in my
implementation.