Create new 'tcb' shell script in place of launchpad
Apparently the internal SSH mechanisms of launchpad only worked on
extremely new OpenSSH versions, which none of the machines accessible to
me had. Without this version keyboard and signal events would not be
forwarded and the UI was not usable...

So I've replaced the launchpad utility with a new shell script that
mostly acomplishes the same thing using rsync. It's named tcb rather
than launchpad because this script only works for call-buddy and doesn't
launch other scripts. For the record, launchpad only did this too but
without the legacy of that, I felt free to name it properly this time.