Go ham on telephono/ to simplify and get state loading/saving working
Pun intended. This commit blows away much of telephono/ in favor of a
less abstracted, more consise and simple approach to environments.
There is no Contributor interface, Expander interface or multiple
implementations of those. There is also no caching; it is not really
necessary and overcomplicates the code. Instead there is a simple
Environment type, which is constructed of a mapping of key to values as
well as a name to go along with the Environment (e.g. User/Var). Setting
a key=value pair is as simple as reaching into CallBuddyEnvironment and
calling the Set() function. This is not the cleanest abstraction, but
it gets the job done* (). It also does this work in
fractions of the original code size within just two files.

So with that hard work (which touched a lot of code), we now have saving
and loading of call-buddy state via Load() and Save(). This is now done
on every call and startup of call-buddy.

Pretty cool :)

* and also I think we need to be realistic about trying to abstract
telephono/ to meet the needs of supposed other users of our module. Doing
so complicates the code and I have my doubts about putting extra work in
code that others could trivially implement.