Implement Ctrl-W keyboard shortcut for Unix machines
This is one of my most used keyboard shortcuts that I use on the
terminal to just delete a word before the cursor. Having it here in
call-buddy just makes the experience even more natural.

This is based off of a new function I added to our custom gocui fork.
Maybe I'll upstream the changes one day.