Use scp to sync up call-buddy binary rather than rsync
scp is known to be somewhat insecure but it is pretty much gauranteed to
be on any remote machine with OpenSSH. On the other hand, I wouldn't be
surprised if rsync wasn't installed on most remote machines (at least
non-interactive machines). So we'll use scp for now for the binary
syncing (where at least the binary is at least not sensitive), but use
rsync for all other things.

I'm not sure if scp is also insecure with regards to data payload
injection (i.e. man in the middle could replace the binary with
something more malicious), but if that is the case we'll have to revert
this commit and insist upon rsync or figure something out with sftp.