Update gocui (again) to fix bug in strict total ordering of Update()s
I think when I upgraded our gocui to use awesome-gocui I didn't apply
the patches I made correctly (or my code was always buggy) so the
result was that the gocui we were just using could deadlock in some
cases after exiting the history view. This has now been fixed in our
gocui with the correct bug fix and the refs here are updated in this
commit.

Who knew that concurrency is hard? ;)