Deny use of tcb command with password-based SSH
We use multiple invocations of SSH-based tools to set up call-buddy in a
remote environment as well as use rsync on a timer to sync new data
down. However, if we allow password-based SSH, the user will have to
enter their password multiple times before even launching into the
remote environment and once there, syncing will not work since they will
not be able to enter their password in.

This is very unfortunate since not all machines you want to connect to
have passwordless SSH set up and setting one up adds a lot of friction.
As far as I'm aware this is by the nature of shell commands and there's
no shell-based alternative that I'm aware of. :(

If I had more time, I'd possibly consider replacing the utility with a
golang based one where I could invoke multiple operations in the same
SSH session, but it's not worth my time at this point.