Don't return our custom errors in evalCmdLine... it causes panic
When GoCui gets an error it doesn't recognize in one of the key handlers
it starts (e.g. onEnter), instead of ignoring it, it panics. (wtf!)

This means we need to be more careful about the errors we return. This
commit makes evalCmdLine more careful about this, though we should
probably make sure other functions are also careful about this in the
future.