Make request header buffer authority for sent headers
We have historically kept a data structure of headers internally that
would be updated with the header command and sent with HTTP requests.
However, this data structure was not kept in sync with the actual
request header view/buffer, making it possible to add a header in the
view and not have it sent!

What we instead want is the buffer itself to be the authority for what
to be send and have the data structure pull from that instead. This is
what this commit does. It introduces an ad-hoc parser for headers that
will update that data structure. It also will warn the user if their
headers are incorrect, though it doesn't handle corner cases like header
names having spaces in them or other weirdness yet.