Avoid blacklisting user specified files in tcb syncdown action
If a user modifies a file they synced up in the remote environment, then
this will not be synced down because the files they give to tcb are
blacklisted for the syncdown action. This is probably not what the user
wants, so I removed the blacklist on user specified files. This does
result in all the files they specified being synced to the local
directory which could be annoying.

Further work to detect what user-specified files have been updated in
the remote target and only sync down those could alleviate this
annoyance.