Abandon go's seemingly broken os/exec.CombinedOutput() function
When you call that function and the command fails, it doesn't output
stderr in the error it returns... in fact, in the description it even
says it doesn't return the entire stderr (it "hold[s] a subset of the
standard error"). I dunno know why this is, it's strange behavior.